Resolving of includes. The path to the file to be included can be either absolute or relative to the directory: 1. Of the main (entry) script 2. Of a currently executed script / the script where the currently executed function or method is defined
To implement this, each program point store information about the file in which it is defined and each function and method declaration (exactly function value) store information about script in which it is defined.

To minimize changes in API, the information about the script which is currently processed are passed to FunctionValue via the static field AnalysisServices.CurrentScript. However, this is bad practise and another solution should be considered.

git-svn-id: svn+ssh://aiya.ms.mff.cuni.cz/var/svn/weverca/trunk@470 67bba403-d36b-46c0-a350-761a85a4c57d