Modified interface of ISnapshotReadWrite.
Added a method ExtendAtSubprogramEntry, which is called at the entry of a subprogram and provides identifiers of all potential callers (there are multiple callers useful when the subprogram is shared). It can be used to keep parts of the state separated for each caller. 

Modified an interface of the method MergeWithCallLevel. Now it identifies the caller of subprograms (can be used to extract parts of the subprograms states corresponding to this particular caller).

TODO: strange thing I have noticed is that ExtensionSinkPoint.extendInput is extended by the caller state (and than CallDispatchMerge is called). I don't see any good reason for extending with the caller state.

git-svn-id: svn+ssh://aiya.ms.mff.cuni.cz/var/svn/weverca/trunk@983 67bba403-d36b-46c0-a350-761a85a4c57d